+++ HISTORY.md
@@ -69,7 +69,8 @@
 * objects: new recordUserEdit() mechanism
 * objects: migrated objects to new recordUserEdit mechanism
 * blocks: migrated blocks to new recordUserEdit mechanism
-* blocks: migrated custo blocks to new recordUserEdit mechanism
+* blocks: migrated custom blocks to new recordUserEdit mechanism
+* threads: migrated processes to new recordUserEdit mechanism
 
 ### 2022-11-26
 * threads: order "my scripts" list by their top-down position in the scripting area